refactor: 발송옵션/질문 패키지 분리 #3
