+++ src/main/java/com/nexters/covid/letter/api/dto/QuestionResponse.java
@@ -2,13 +2,15 @@ package com.nexters.covid.letter.api.dto;
 
 import static org.springframework.beans.BeanUtils.copyProperties;
 
-import com.nexters.covid.letter.domain.Question;
+import com.nexters.covid.letter.domain.question.Question;
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
 public class QuestionResponse {
+
+  private Long id;
 
   private String text;
