+++ src/main/java/com/nexters/covid/letter/service/LetterService.java
@@ -6,8 +6,8 @@ import com.nexters.covid.letter.api.dto.OptionResponse;
 import com.nexters.covid.letter.api.dto.QuestionResponse;
 import com.nexters.covid.letter.domain.Letter;
 import com.nexters.covid.letter.domain.LetterRepository;
-import com.nexters.covid.letter.domain.QuestionRepository;
-import com.nexters.covid.letter.domain.SendOptionRepository;
+import com.nexters.covid.letter.domain.question.QuestionRepository;
+import com.nexters.covid.letter.domain.sendoption.SendOptionRepository;
 import com.nexters.covid.user.domain.User;
 import com.nexters.covid.user.domain.UserRepository;
 import java.util.List;
